Refactoring models: schema.org in submodule
