+++ cads_catalogue_api_service/models/__init__.py
@@ -1,0 +1,23 @@
+from . import schema_org
+from .base import (
+    Changelog,
+    Keyword,
+    Keywords,
+    Licence,
+    LicenceCategories,
+    Licences,
+    Message,
+    Messages,
+)
+
+__all__ = [
+    "LicenceCategories",
+    "Licence",
+    "Licences",
+    "Keyword",
+    "Keywords",
+    "Message",
+    "Messages",
+    "Changelog",
+    "schema_org",
+]
