+++ cads_catalogue_api_service/models/schema_org.py
@@ -1,6 +1,6 @@
-"""Custom models that extends the OGC/STAC ones."""
+"""Custom Schema.org (Google Datasets compatibility)."""
 
-# Copyright 2022, European Union.
+# Copyright 2023, European Union.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,107 +14,43 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import datetime
-import enum
-from typing import TypedDict
-
 import pydantic
 
 
-class LicenceCategories(str, enum.Enum):
-    dataset: str = "dataset"
-    portal: str = "portal"
-
-
-class Licence(TypedDict):
-    """Licence definition."""
-
-    # id here is the licence slug
-    id: str
-    label: str
-    revision: int
-    contents_url: str
-    attachment_url: str
-    scope: LicenceCategories
-
-
-class Licences(TypedDict):
-    """Licences vocabulary."""
-
-    licences: list[Licence]
-
-
-class Keyword(TypedDict):
-    """Keyword definition."""
-
-    id: str
-    label: str
-
-
-class Keywords(TypedDict):
-    """Keywords vocabulary."""
-
-    keywords: list[Keyword]
-
-
-class Message(TypedDict):
-    """Message definition."""
-
-    id: str
-    date: datetime.datetime
-    summary: str | None
-    url: str | None
-    severity: str
-    content: str
-    live: bool
-
-
-class Messages(TypedDict):
-    """Messages vocabulary."""
-
-    messages: list[Message]
-
-
-class Changelog(TypedDict):
-    """Changelog vocabulary."""
-
-    changelog: list[Message]
-
-
-class SchemaOrgContactPoint(pydantic.BaseModel):
+class ContactPoint(pydantic.BaseModel):
     type: str = pydantic.Field("ContactPoint", const=True, alias="@type")
     contact_type: str = pydantic.Field(alias="contactType")
     email: str
     url: str
 
 
-class SchemaOrgOrganization(pydantic.BaseModel):
+class Organization(pydantic.BaseModel):
     type: str = pydantic.Field("Organization", const=True, alias="@type")
     url: str
     name: str
     logo: str | None = None
-    contact_point: SchemaOrgContactPoint | None = pydantic.Field(
+    contact_point: ContactPoint | None = pydantic.Field(
         ..., serialization_alias="contactPoint"
     )
 
 
-class SchemaOrgDataDownload(pydantic.BaseModel):
+class DataDownload(pydantic.BaseModel):
     type: str = pydantic.Field("DataDownload", const=True, alias="@type")
     encoding_format: str = pydantic.Field(alias="encodingFormat")
     content_url: str = pydantic.Field(alias="contentUrl")
 
 
-class SchemaOrgGeoShape(pydantic.BaseModel):
+class GeoShape(pydantic.BaseModel):
     type: str = pydantic.Field("GeoShape", const=True, alias="@type")
     box: list[float]
 
 
-class SchemaOrgPlace(pydantic.BaseModel):
+class Place(pydantic.BaseModel):
     type: str = pydantic.Field("Place", const=True, alias="@type")
-    geo: SchemaOrgGeoShape
+    geo: GeoShape
 
 
-class SchemaOrgDataset(pydantic.BaseModel):
+class Dataset(pydantic.BaseModel):
     context: str = pydantic.Field("https://schema.org/", const=True, alias="@context")
     type: str = pydantic.Field("Dataset", const=True, alias="@type")
 
@@ -128,9 +64,9 @@ class SchemaOrgDataset(pydantic.BaseModel):
     is_accessible_for_free: bool = pydantic.Field(
         alias="asAccessibleForFree", default=True
     )
-    creator: SchemaOrgOrganization
-    distribution: list[SchemaOrgDataDownload]
+    creator: Organization
+    distribution: list[DataDownload]
     temporalCoverage: str | None
-    spatialCoverage: SchemaOrgPlace
+    spatialCoverage: Place
     dateModified: str | None
     thumbnailUrl: str | None = None
