+++ cads_catalogue_api_service/schema_org.py
@@ -25,116 +25,6 @@ from cads_catalogue_api_service.client import collection_serializer
 
 from . import dependencies, models
 
-TEMPLATE = {
-    "@context": "https://schema.org/",
-    "@type": "Dataset",
-    "name": "NCDC Storm Events Database",
-    "description": """Storm Data is provided by the National Weather
-    Service (NWS) and contain statistics on...""",
-    "url": "https://catalog.data.gov/dataset/ncdc-storm-events-database",
-    "sameAs": "https://gis.ncdc.noaa.gov/geoportal/catalog/search/resource/details.page?id=gov.noaa.ncdc:C00510",
-    "identifier": [
-        "https://doi.org/10.1000/182",
-        "https://identifiers.org/ark:/12345/fk1234",
-    ],
-    "keywords": [
-        "ATMOSPHERE > ATMOSPHERIC PHENOMENA > CYCLONES",
-        "ATMOSPHERE > ATMOSPHERIC PHENOMENA > DROUGHT",
-        "ATMOSPHERE > ATMOSPHERIC PHENOMENA > FOG",
-        "ATMOSPHERE > ATMOSPHERIC PHENOMENA > FREEZE",
-    ],
-    # For some datasets we have multiple licences, so we'll use the
-    #  STAC way of reportimg them (AKA: "various")
-    "license": "https://creativecommons.org/publicdomain/zero/1.0/",
-    "isAccessibleForFree": True,
-    "creator": {
-        "@type": "Organization",
-        "url": "https://climate.copernicus.eu",
-        "name": "Copernicus Climate Change Service",
-        "logo": "https://climate.copernicus.eu/themes/custom/ce/logo.svg",
-        "contactPoint": {
-            "@type": "ContactPoint",
-            "contactType": "customer service",
-            "email": "copernicus-support@ecmwf.int",
-            "url": "https://cds.climate.copernicus.eu/contact-us",
-        },
-    },
-    "funder": {
-        "@type": "Organization",
-        "sameAs": "https://ror.org/00tgqzw13",
-        "name": "National Weather Service",
-    },
-    "includedInDataCatalog": {"@type": "DataCatalog", "name": "data.gov"},
-    "distribution": [
-        {
-            "@type": "DataDownload",
-            "encodingFormat": "CSV",
-            "contentUrl": "https://www.ncdc.noaa.gov/stormevents/ftp.jsp",
-        },
-        {
-            "@type": "DataDownload",
-            "encodingFormat": "XML",
-            "contentUrl": "https://gis.ncdc.noaa.gov/all-records/catalog/search/resource/details.page?id=gov.noaa.ncdc:C00510",
-        },
-    ],
-    "temporalCoverage": "1950-01-01/2013-12-18",
-    "spatialCoverage": {
-        "@type": "Place",
-        "geo": {"@type": "GeoShape", "box": "18.0 -65.0 72.0 172.0"},
-    },
-    # "variableMeasured":
-    # "dateModified":
-    # "discussionUrl":
-    # "abstract":
-    # "thumbnail" o "thumbnailUrl":
-    # "text": ???
-}
-
-
-ECMWF_TEMPLATE = {
-    "@context": "http://schema.org/",
-    "@type": "Dataset",
-    "name": "Fire danger indices historical data from the Copernicus Emergency Management Service",
-    "description": """This data set provides complete historical reconstruction of meteorological
-    conditions favourable to the start, spread and sustainability of fires. The fire danger metrics
-    provided are part of a vast dataset produced by the Copernicus Emergency Management Service for
-    the European Forest Fire Information System (EFFIS). The European Forest Fire Information System
-    incorporates the fire danger indices for three different models developed in Canada, United States
-    and Australia. In this dataset the fire danger indices are calculated using weather forecast from
-    historical simulations provided by ECMWF ERA5 reanalysis.
-    ERA5 by combining  model data and  a vast set of quality controlled observations provides a  globally
-    complete and consistent data-set and is regarded as a good proxy for observed atmospheric conditions.
-    The selected data records in this data set are regularly extended with time as ERA5 forcing data become
-     available.
-    This dataset is produced by ECMWF in its role of the computational centre for fire danger forecast of
-    the CEMS,  on behalf of the Joint Research Centre which is the managing entity of the service.""",
-    "url": "https://cds.climate.copernicus.eu/cdsapp#!/dataset/cems-fire-historical-v1",
-    "sameAs": "https://cds.climate.copernicus.eu/dataset/cems-fire-historical-v1",
-    "license": "licence-to-use-copernicus-products",
-    "keywords": ["PRODUCT TYPE >  REANALYSIS", "SPATIAL COVERAGE >  GLOBAL"],
-    "creator": {
-        "@type": "Organization",
-        "url": "https://climate.copernicus.eu",
-        "name": "Copernicus Climate Change Service",
-        "logo": "https://climate.copernicus.eu/themes/custom/ce/logo.svg",
-        "contactPoint": {
-            "@type": "ContactPoint",
-            "contactType": "customer service",
-            "email": "copernicus-support@ecmwf.int",
-            "url": "https://cds.climate.copernicus.eu/contact-us",
-        },
-    },
-    "distribution": [
-        {
-            "@type": "DataDownload",
-            "encodingFormat": "NetCDF (interpolated grid only)  and GRIB2 (all grids)",
-            "contentUrl": "https://cds.climate.copernicus.eu/cdsapp#!/dataset/cems-fire-historical-v1",
-        }
-    ],
-    "temporalCoverage": "Daily",
-}
-
-
 router = fastapi.APIRouter(
     prefix="",
     tags=["messages"],
@@ -166,13 +56,13 @@ def get_url_link(collection: stac_fastapi.types.stac.Collection, type: str) -> s
 
 @router.get(
     "/collections/{collection_id}/schema.org",
-    response_model=models.SchemaOrgDataset,
+    response_model=models.schema_org.Dataset,
 )
 def schema_org_jsonId(
     collection_id: str,
     request: fastapi.Request,
     session=fastapi.Depends(dependencies.get_session),
-) -> models.SchemaOrgDataset:
+) -> models.schema_org.Dataset:
     """Endpoint to get the proper schema.org compatible definition for the dataset.
 
     See https://developers.google.com/search/docs/appearance/structured-data/dataset
@@ -192,7 +82,7 @@ def schema_org_jsonId(
 
     box = collection.get("extent", {}).get("spatial", {}).get("bbox", [])
 
-    return models.SchemaOrgDataset(
+    return models.schema_org.Dataset(
         context="http://schema.org/",
         type="Dataset",
         name=collection["title"],
@@ -204,10 +94,10 @@ def schema_org_jsonId(
         license=license,
         keywords=collection.get("keywords", []),
         is_accessible_for_free=True,
-        creator=models.SchemaOrgOrganization(
+        creator=models.schema_org.Organization(
             url=collection.get("creator_url", None),
             name=collection["creator_name"],
-            contact_point=models.SchemaOrgContactPoint(
+            contact_point=models.schema_org.ContactPoint(
                 contactType="User support",
                 # FIXME: This is a problem with input data
                 email=collection.get("creator_contact_email", None),
@@ -215,7 +105,7 @@ def schema_org_jsonId(
             ),
         ),
         distribution=[
-            models.SchemaOrgDataDownload(
+            models.schema_org.DataDownload(
                 encodingFormat=collection.get("file_format"),
                 contentUrl=f"{url}?tab=download",
             )
@@ -223,9 +113,9 @@ def schema_org_jsonId(
             else ""
         ],
         temporalCoverage="/".join(temporal_coverage) if temporal_coverage else None,
-        spatialCoverage=models.SchemaOrgPlace(
+        spatialCoverage=models.schema_org.Place(
             type="",
-            geo=models.SchemaOrgGeoShape(
+            geo=models.schema_org.GeoShape(
                 type="",
                 box=box[0] if box else [],
             ),
